requirements.txt
google-genai>=1.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
